// types.ts
export type UserRole = 'provider' | 'receiver' | 'volunteer' | 'admin_manager' | 'super_admin' | null;
// ...
export interface Address {
  id: string;
  userId?: string;
  role?: UserRole;
  label: string;
  fullAddress: string;
  receiverName: string;
  phone: string;
  isPrimary: boolean;
  lat?: number;
  lng?: number;
}